refactor to put CollectionCase and UAC into model package, other classes into message package, and delete CaseAddress class
