+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/message/CaseEventReceiver.java
@@ -1,7 +1,6 @@
 package uk.gov.ons.ctp.integration.rhsvc.message;
 
 import org.springframework.integration.annotation.MessageEndpoint;
-import uk.gov.ons.ctp.integration.rhsvc.message.impl.CaseEvent;
 
 /**
  * Service implementation responsible for receipt of Case Events. See Spring Integration flow for
