+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/message/CasePayload.java
@@ -1,9 +1,10 @@
-package uk.gov.ons.ctp.integration.rhsvc.message.impl;
+package uk.gov.ons.ctp.integration.rhsvc.message;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import uk.gov.ons.ctp.integration.rhsvc.domain.model.CollectionCase;
 
 @Data
 @Builder
