+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/message/GenericEvent.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.rhsvc.message.impl;
+package uk.gov.ons.ctp.integration.rhsvc.message;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
