+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/message/Header.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.rhsvc.message.impl;
+package uk.gov.ons.ctp.integration.rhsvc.message;
 
 import java.sql.Timestamp;
 import lombok.AllArgsConstructor;
