+++ src/test/java/uk/gov/ons/ctp/integration/rhsvc/service/impl/RespondentDataServiceIT_Test.java
@@ -6,8 +6,8 @@ import java.util.Optional;
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
-import uk.gov.ons.ctp.integration.rhsvc.message.impl.CollectionCase;
-import uk.gov.ons.ctp.integration.rhsvc.message.impl.UAC;
+import uk.gov.ons.ctp.integration.rhsvc.domain.model.CollectionCase;
+import uk.gov.ons.ctp.integration.rhsvc.domain.model.UAC;
 import uk.gov.ons.ctp.integration.rhsvc.service.RespondentDataService;
 
 // import uk.gov.ons.ctp.integration.rhsvc.domain.model.Address;
